+++ evaluation/src/scores.rs
@@ -1,4 +1,4 @@
-const SCORE_INF: i32 = 30_000;
-pub const POS_INFINITY: i32 = SCORE_INF;
-pub const NEG_INFINITY: i32 = -SCORE_INF;
-pub const MATE_VALUE: i32 = SCORE_INF - 1000;
+const SCORE_INF: i16 = 30_000;
+pub const POS_INFINITY: i16 = SCORE_INF;
+pub const NEG_INFINITY: i16 = -SCORE_INF;
+pub const MATE_VALUE: i16 = SCORE_INF - 1000;
